// src/App.d.ts
import './App.css';
export interface Currency {
    symbol: string;
    priceChange: string;
    priceChangePercent: string;
    weightedAvgPrice: string;
    prevClosePrice: string;
    lastPrice: string;
    lastQty: string;
    bidPrice: string;
    bidQty: string;
    askPrice: string;
    askQty: string;
    openPrice: string;
    highPrice: string;
    lowPrice: string;
    volume: string;
    quoteVolume: string;
    openTime: number;
    closeTime: number;
    firstId: number;
    lastId: number;
    count: number;
}
export interface Term {
    [x: string]: {
        symbol: string;
        countTerm: number;
    };
}
declare function App(): import("react/jsx-runtime").JSX.Element;
export default App;